DogF4 Whale Mode Split + cleanup and re-updated code
